Removing transaction
The 'INCR' command wasn't returning a real value until the
transaction was reaching 'EXEC'. But I was expecting an integer
for the next ID before then. So the 'DEFER' value returned by 'INCR'
was not an int, and there was an error because turning 'DEFER' into
`nextID` was obviously failing.

Two lessons: debugger in VS Code and `log.Printf` are handy. And
*always* use the `err` value. Bonus: use a return value in error cases
that is not the same as a default value in Go. My return value of zero
in err cases is the same as the default value for ints in Go. So
I thought that the zero indicated an error in Unmarshalling the object
to JSON, but actually it's because I was explicitly setting
the value to zero in error catch blocks. Now its' -1 to be super
obvious.